Added a basic connection mechanism for server and client
A server will accept only one client, and inmediately it will be closed
when a client connects to it. A client could connect to a predefined IP.
If it is not defined, program will ask user for a IP while there is not
connection or user doesn't type anything. Default port will be 3856.
Server will use a network interface connected to internet. For example,
wlp2s0 or eth0.